Integration of Jukebox Library (#451)
* add Jukebox library

todo:
- saving / loading of new format
- inferrence of unknown data like past versions
- move the temporary float casts to proper use of double
- make sound related functions take double for timing
- inform people that the Jukebox sound player was renamed to SoundByte lol

* make sound, input scheduling, and super curve use double precision

* successfully load charts

* editor works again

v1 riqs can be saved and loaded

* first tempo and volume markers are unmovable

fix loading of charts' easing values

* use gsync / freesync

* update Jukebox refs to SoundByte

* game events use double part 1

Air Rally - Glee Club converted

* don't load song if chart load fails

* finish conversion of all minigames

* remove editor waveform toggle

* timeline now respects added song offset length

clear cache files on app close
prepped notes for dsp sync

* update timeline length when offset changed

* update to latest Jukebox

* make error panel object in global game manager

* improve conductor music scheduling

* added error message box

fix first game events sometimes not playing

+++ Assets/Scripts/Games/DoubleDate/DoubleDateWeasels.cs
@@ -42,7 +42,7 @@ namespace HeavenStudio.Games.Scripts_DoubleDate
             }
         }
 
-        public void Hide(float beat)
+        public void Hide(double beat)
         {
             if (notHit)
             {
@@ -69,7 +69,7 @@ namespace HeavenStudio.Games.Scripts_DoubleDate
             }
         }
 
-        public void Hit(float beat)
+        public void Hit(double beat)
         {
             if (notHit)
             {
